move start related functions
